+++ CHANGELOG.rst
@@ -9,6 +9,8 @@
   not just closing the connection
 * `#512 <https://github.com/xitrum-framework/xitrum/issues/512>`_
   Add config maxHeaderSize (for the entire HTTP header, defaults to 8192)
+* `#518 <https://github.com/xitrum-framework/xitrum/issues/518>`_
+  xitrum.util.SeriDeseri.fromXXX[T] defaults to String
 * `#517 <https://github.com/xitrum-framework/xitrum/issues/517>`_
   Update Scala to 2.10.5 and 2.11.6
 * `#511 <https://github.com/xitrum-framework/xitrum/issues/511>`_